Extract Constraint handling into AbstractConstraintLayout
XYLayout and RulerLayout share quite some common code. This is extracted
in a common base class to be used also by other Layouts.
